Final
In this commit I read over the rubric and corrected/added things that I have missed. This includes <abbr> for BC and PHD, etc. And also fixed a date-time for machine reading. Also used the sub and sup for the chemical formulas and mathematical expressions. Removed the base structure with my plan in the final copy.